Remove StringExtensions and UInt64Extensions files
The `StringExtensions.cs` file, which contained the `ToLong` method for converting a string to an unsigned long (ulong), has been removed. Similarly, the `UInt64Extensions.cs` file, which contained the `ToBoolean` method for converting an unsigned long (ulong) to a boolean, has also been removed. These extension methods are no longer part of the codebase.